iio: adc: qcom-spmi-adc5-gen3: update TM error handling logic
At present, if there is a conversion fault interrupt for a TM conversion
channel, we attempt to clear the fault and immediately reconfigure the
TM conversion in the interrupt handler by calling adc_tm5_gen3_configure.

However, adc_tm5_gen3_configure internally calls adc5_gen3_poll_wait_hs
which in turn calls usleep_range, which is a sleeping function and
should not be called in interrupt context.

In addition, before adc_tm5_gen3_configure is called, the adc mutex_lock
needs to be taken as adc_tm5_gen3_configure writes to registers which
may be used for either immediate or TM conversions, to avoid potential
race conditions.

Therefore, add a new work struct under adc5_chip to reconfigure TM
conversions and queue this work when TM conversion fault is detected in the
interrupt handler.

Fixes: e1dc4cf289f2 ("iio: adc: qcom-spmi-adc5-gen3: Reconfigure ADC_TM channel after fault")
Change-Id: Id53ad0ba8fcb92a4a2d9d50a9304636e8b659f96
Signed-off-by: Jishnu Prakash <quic_jprakash@quicinc.com>