sched/walt: Fix issue in RT packing logic
It could be unsafe to access the current task on a random runqueue
without holding locks. Furthermore, the check for looking at the current
task was initially introduced and meant only to be done on the waking
cpu, thereby limiting RT packing to just one task per CPU. However, due
to the incorrect way in which it was constructed, the behavior
ultimately ended up resulting in up to 2 tasks being allowed to be
packed per CPU.

To simplify things, allow RT packing of up to two tasks per cpu, and
remove checking for the policy of the task of the packing cpu.

Change-Id: I2b88a5a69c972cae4816cb088dcbde23d37b8cd1
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>