mem-buf-gh: alloc altmap for large dmabuf only
The commit ("mem-buf-gh: support large dmabuf memory lending")
was doing remote alloc for memmap memory for dmabuf that are
equal or larger than 128MB and it would also do so if the
zone watermarks are not ok when memmap memory will be allocated
from Normal zone. The later check would now allow dmabuf of
any size to have its memmap memory be allocated from PVM if
zone watermarks are not ok. This was causing the below issue:
	Unable to handle kernel paging request at virtual address fffffffe01808000
	Mem abort info:
	pc : memunmap_pages+0x1a4/0x2a8
	lr : memunmap_pages+0x1a4/0x2a8
	Call trace:
	memunmap_pages+0x1a4/0x2a8
	mem_buf_retrieve_dma_buf_release+0x20/0x70
	dma_buf_release+0x40/0x9c
	__dentry_kill+0x17c/0x1ec
	dentry_kill+0xbc/0x104
	dput+0xc4/0x128
	__fput+0x180/0x288
	__fput_sync+0x28/0x60
	__arm64_sys_close+0x38/0x74
	invoke_syscall+0x40/0xfc
Until this issue is root caused properly, remove the watermark
check so that we only do remote alloc for memmap for dmabuf
of size >= 128MB.

Change-Id: I3bb77e41f855dae17609c4d1f444854505b6113e
Signed-off-by: Sudarshan Rajagopalan <quic_sudaraja@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>