DNM: cpufreq: scmi: Assign target_freq to u64 to avoid possible overflow
The target_freq variable in scmi_cpufreq_fast_switch() is unsigned int
currently, and target_freq is converted to Hz and passed as an input to
scmi perf .freq_set() callback. But this conversion can lead to overflow
for large values of target_freq, avoid this overflow issue by assigning
target_freq to u64 variable for converting it to Hz.

Signed-off-by: Jagadeesh Kona <quic_jkona@quicinc.com>
Change-Id: I2c1e62ff3af35b64ec759b9cc1606fc392e1709f