DNM: cpufreq: scmi: Assign target_freq to u64 to avoid possible overflow
The target_freq variable in scmi_cpufreq_fast_switch() is unsigned int
currently, and target_freq is converted to Hz and passed as an input to
scmi perf .freq_set() callback. But this conversion can lead to overflow
for large values of target_freq, avoid this overflow issue by assigning
target_freq to u64 variable for converting it to Hz.

Change-Id: I0499dac86557ac049a434af81461c3a2d2cfe2ae
Signed-off-by: Jagadeesh Kona <quic_jkona@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>