sound: usb: qmi: Add NULL check for chip->intf[0]
Consider a case where driver went ahead to enable the audio stream
and tries to call pm_runtime_barrier on xhci to make sure that xhci
is resumed properly. Here there could be a possibility of chip->intf[0]
being NULL and the driver is trying to access the device of that
particular interface.

Fix this by adding a check for chip->intf[0] before accessing it.

Change-Id: I506fab5b6edca08cbc0dd570f8aab055f3d0ed40
Signed-off-by: Pratham Pratap <quic_ppratap@quicinc.com>