usb: dwc3: dwc3-msm-ops: Stop pending GSI transfers on reset
In order to ensure that GSI transfers are stopped on a bus reset condition,
do not rely on f_gsi issuing the EP op to do it.  f_gsi would handle this
asynchronously within the IPA SM work, which could run later in the
timeline when the bus reset routine has already completed.  Utilize the
DWC3 reset kretprobe to issue the endxfer command for any pending GSI
endpoints.

Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>
Change-Id: I6d973a839095c335e845ba46d9c05b43bdf33a60