firmware: scm: Avoid NULL pointer dereference issue
caller of qcom_scm_set_boot_addr() incorrectly calls SMC
api without checking if SCM driver is ready or not without
which it is possible it can fail in low level SCM layer
and result in NULL pointer dereference.

While at it, also fix the place of scm device assignment.

Change-Id: Ifd733ee84235932e1ad001bc0e07cfa3f255dfe7
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>