dmaengine: gpi: Clean up the GSI HW after VM reboot
Primary VM needs memory and hence if finds that VM is inactive or
no usecase is running on VM, it can trigger shutdown of VM and get
the memory freed. After VM reboot, priorer to first gsi transfer
checking if GSI HW is already initialized, if so clearing/resetting
the GSI HW.

In this way putting GSI HW to default state after every VM reboot.

Change-Id: I8b48a4996929048c4f4c6eaffe80f8e9cfdd94b9
Signed-off-by: Anil Veshala Veshala <quic_aveshala@quicinc.com>
Signed-off-by: Krishna Chaithanya Reddy G <quic_kgangapu@quicinc.com>