sched/walt: Fix issue with setting boost for OBET
sbt_ctl_check() was extended to handle the need to unisolate Large
cluster cpus while in OBET. However the implementation was incorrect as
it ends up calling the core ctl boost api everytime it is invoked,
even if obet didn't change.

This causes the boost counters to infinitely increase or decrease
effectively disabling other boost reasons.

Fix this by ensuring the boost is called only if obet signal has
changed.

Change-Id: I6623da2454070f99752e5f3c228e0dafbbe0085c
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>