drivers: dcvs: bwmon: Add cpumask to low power io percent
The initial implementation of low power io percent only
allows for tracking a single cpu/cluster's frequency.
Improve this by adding support to track the frequency of
all clusters in the system. This allows the low power io
percent feature to only trigger when all clusters are
running at low enough frequencies. To ignore a cluster,
user can just set a high value and to disable the
feature, any cluster max freq can be set to 0.

Change-Id: I74b6aaa20d3e8e7f7d35f10e5d5ff0afe778dcea
Signed-off-by: Amir Vajid <quic_avajid@quicinc.com>