clk: qcom: dispcc-sun: Keep mdss_mdp_clk always enabled
The disp_cc_mdss_mdp_clk is required for its parent to disable properly.
The parent clock (disp_cc_mdss_mdp_clk_src) is managed by a CESTA VCD.
When that VCD collapses, it triggers a handshake with display HW that
requires this clock to complete. Thus, keep the clock enabled such that
it's not disabled before its parent. The clock will eventually turn off
when the disp_cc_mdss_core_gdsc is disabled.

Change-Id: Id4b554272a0e40e0b46e0b6d54ee62ea446df924
Signed-off-by: Mike Tipton <quic_mdtipton@quicinc.com>