remoteproc: qcom: pas: Unlock metadata after auth and reset
mdt loader for remoteprocs that do not free PAS metadata that is
passed to the secure world for SoC where TME based authentication
is done e.g., modem. This is required as the secure world
locks the metadata buffer to ensure that no modifications to
the firmware can be made. The metadata buffer is unlocked once
the auth and reset call succeeds, after which the HLOS can free
the metadata and if any intermediate call till auth and reset
fails then it need to call shutdown call to unlock the metadata
region however, this is not true for ADSP and CDSP where meta
data are unlocked and ready to released right after
qcom_scm_pas_init_image() call.

c64665ac0a17 ("remoteproc: pas: Release dtb CMA memory after SCM
 shutdown call on dtb pas id") was not completely fix the issue.
This attempt is to fix the issue completely.

Fixes: c64665ac0a17 ("remoteproc: pas: Release dtb CMA memory after SCM shutdown call on dtb pas id")
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>
Change-Id: Idffd0578b24b3124e2540ac06d1707513f2607b4