remoteproc: qcom: pas: Unlock metadata after auth and reset
mdt loader for remoteprocs that do not free PAS metadata that is
passed to the secure world for SoC where TME based authentication
is done e.g., modem. This is required as the secure world
locks the metadata buffer to ensure that no modifications to
the firmware can be made. The metadata buffer is unlocked once
the auth and reset call succeeds, after which the HLOS can free
the metadata and if any intermediate call till auth and reset
fails then it need to call shutdown call to unlock the metadata
region however, this is not true for ADSP and CDSP where metadata
are unlocked and ready to released right after qcom_scm_pas_init_image()
call.

commit c64665ac0a17 ("remoteproc: pas: Release dtb CMA memory after
 SCM shutdown call on dtb pas id") was doing the clean up for subsystem's
dtb metadata while we should also clean up subsystem core image metadata
as well.

Fixes: c64665ac0a17 ("remoteproc: pas: Release dtb CMA memory after SCM shutdown call on dtb pas id")
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>
Change-Id: I011f42e952f081b256fbbf9ea0dfb89f900cfb8a