drivers: dcvs: bwmon: Fix secondary AB vote calculation
The bwmon secondary voter directly uses the primary AB
vote in units of kHz. This is not correct since the
secondary voter may have a different bus width. Use the
primary AB vote in units of MBPS when calculating the
secondary AB vote to fix this. Also update the AB scale
tunable max to 200 to allow for scaling up AB in tuning.

Change-Id: I8ab4baedd80715c8abe58bca29b972eae8aafdda
Signed-off-by: Amir Vajid <quic_avajid@quicinc.com>