sched: walt: revise the criterion for task misfit detection
A task should only be classified as a misfit if it has the
capability to run on more than one cluster. In other words,
if a task is affined solely to the CPUs of a single cluster,
it should be considered a suitable fit for the CPUs within
that cluster.

Change-Id: Ibbb13a7e310f81944e87e3de430792065586290f
Signed-off-by: Ashay Jaiswal <quic_ashayj@quicinc.com>