usb: dwc3-msm-core: Prevent unwanted panic in pwr_event_handler
Currently, the function pwr_event_handler check only for a valid
dwc and proceeds, however there are operations w.r.t power event
irq which are still accessible and if use_pwr_event_for_wakeup
isn't set, those settings should not be exercised.
Also, even if the dwc is NULL the rest of the functions can still
be executed which is not correct.

Fix this by correcting the early bailout condition.

Change-Id: Ie4dd127f02399ec3b4968c660067c0d0c486e8ad
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>