qti_virtio_mem: Support plugging in memory at probe
"qcom,initial-movable-zone-size" creates a movable zone of the given size
before userspace starts. This is useful in a system where a userspace
component (psi-daemon) is used to hotplug in memory on a on-demand basis.

With swap disabled, the current tvm carveout size (68 Mb) hits
out-of-memory condition before psi-deamon is started.

Once psi-daemon is running, it detects the number of blocks added to the
movable zone via "qti_virtio_mem/kernel_plugged" sysfs file, and
can reduce or increment the number as necessary.

Change-Id: I2d4c8b0b2fd7dda527d2217cf9fedb236aaea216
Signed-off-by: Patrick Daly <quic_pdaly@quicinc.com>