sched/walt: Fix IPC freq update after tickless mode
Consider a sitution on a CPU where the IPC level is 4 (IPC E). When the
CPU enters tickless mode, the IPC level is reset to 0. However, if the
IPC level remains at 4 (IPC E) when the CPU receives the next tick,
the update to the CPU frequency governor is skipped, as there has been
no change between the current and the last IPC level. This results in
the frequency allowed being kept at level 0, even though the CPU
has undergone a tick.

To address this issue, introduce a per-CPU variable to mark the
transition into tickless mode. This can be used to ensure that as soon
as the CPU receives a tick, there will be an update to the governor,
regardless of the current IPC level.

Change-Id: I009588c8095e687c35dbe3ab35a7754a5e9bc9d6
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>