sched: walt: add policy to handle frequent yielding task
Add policy in WALT to handle task reqlinquishing cpu frequently.
If a task does sched_yield more that a threshold count in a
wakeup-run then WALT treats that task as a frequent yielder task.
Once a task is marked as frequent yielder following policy is applied
for the task:
For all the subsequent yields task, a sleep is induced and task
sleeps instead of yielding cpu and staying on runqueue.
The 'yield count' of a task is reset back to 0 under following
condition:
- If task sleeps naturally(not yield induced sleep).
- Every window rollover if task has not crossed the threshold.

Change-Id: Ie40da6976b27868d879d8fd0aea5ff5189a7fbed
Signed-off-by: Ashay Jaiswal <quic_ashayj@quicinc.com>