sound: usb: usb_audio_qmi_svc: Set QMI return status correctly on error
In some of the exit early conditions, the ret value is not overridden with
an error code.  This would lead to a condition where the QMI response
status would be zero (success), while other fields are not properly
populated.  The error condition in question is:

uaudio_err("invalid substream\n");

Address this by setting the ret value manually in case of an error
condition.  This would allow for the audio DSP to treat the QMI response as
a failure versus attempting to continue processing it.  The issue was
seen if the audio device is disconnected shortly after attempting to
start the audio session.

Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>
Change-Id: I89326e6f082109176ed1647df45dd9021cf663af