sched/walt: runtime bitmap for tasks
Currently walt reacts to a cpufreq normalized view of the task load for
demand and cpufreq updates.

While this is adequate for most of the usecases, certain critical
display usecases need to know the precise runtime of tasks.
Create a bitmap representation of the runtime, such that every bit in
that bitmap represents 1mS and a set bit in that bitmap indicates that
the task has run for majority of time in that mS.

This will be used in upcoming patches for boosting cpufreq.

Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>
Signed-off-by: Abhijeet Dharmapurikar <quic_adharmap@quicinc.com>
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>
Change-Id: Icc05799dbb220a770e39f5b4e3c797413e08e331