mem-buf-gh: align memmap_base to section size boundary
Use the gen_pool_alloc_algo variant for allocating memory from
the genpool with an alignment requirement.

Previously, for the altmap creation, the memmap_base that we
would get from gen_pool_alloc would not be aligned to memory
block boundary (i.e. 128MB for ARM64 targets), which leads to
scenario where a memmap_base would be part of an existing memory
block that was already created and exists. This would lead to
partial initialization of the altmap, which then leads to the
below crash in the system -

[17737.825632][ T2457] Unable to handle kernel paging request at virtual address fffffffe01a48000
[17737.825673][ T2457] swapper pgtable: 4k pages, 39-bit VAs, pgdp=00000000890ae000
[17737.825674][ T2457] [fffffffe01a48000] pgd=100000008cb22003, p4d=100000008cb22003, pud=100000008cb22003, pmd=0000000000000000
[17737.825708][ T2457] pc : memunmap_pages+0x1a4/0x2a8
[17737.825714][ T2457] lr : memunmap_pages+0x1a4/0x2a8
[17737.825742][ T2457] Call trace:
[17737.825743][ T2457]  memunmap_pages+0x1a4/0x2a8
[17737.825744][ T2457]  mem_buf_retrieve_dma_buf_release+0x20/0x70
[17737.825748][ T2457]  dma_buf_release+0x40/0x9c
[17737.825751][ T2457]  __dentry_kill+0x17c/0x1ec
[17737.825754][ T2457]  dentry_kill+0xbc/0x104
[17737.825756][ T2457]  dput+0xc4/0x128
[17737.825758][ T2457]  __fput+0x180/0x288
[17737.825760][ T2457]  __fput_sync+0x28/0x60
[17737.825763][ T2457]  __arm64_sys_close+0x38/0x74
[17737.825765][ T2457]  invoke_syscall+0x40/0xfc

By aligning the memmap_base to memory block boundary resolves
this issue. Do this by setting min_alloc_order = section size
order during gen pool creation.

Change-Id: Ic60f6ea273d8543951600f59aea1b9f5c1eb7a0f
Fixes: 0404725a777a ("mem-buf-gh: support large dmabuf memory lending")
Signed-off-by: Sudarshan Rajagopalan <quic_sudaraja@quicinc.com>